tests: Generate KeyFileExt with KeyFile
Currently, the writekeyfile.c program generates an example KeyFile that
we use for the auth/keys-t tests. But this only contains the old-style
single-DES keys, which are not recommended and not as common anymore.

To test our handling of the KeyFileExt format with newer encryption, add
some tests in auth/keys-t to handle existing afsconf_rxkad_krb5 keys in
KeyFileExt. Change writekeyfile.c to also write out a few
afsconf_rxkad_krb5 keys in a sample KeyFileExt file.

To do this, rewrite writekeyfile.c to make it easier to add new
generated keys and files, and rename it to genkeyfiles.c. Add the
generated KeyFileExt, and modify auth/writeoldkey-t to test that the
program generates the correct generated KeyFileExt file (like we did for
KeyFile). Rename auth/writeoldkey-t to auth/genkeyfiles-t to more
accurately reflect what it is testing.

Change-Id: I94188acdc7ad6dc1b4fc713d6b278044f6c87a1b
Reviewed-on: https://gerrit.openafs.org/16398
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>