tests: Add rxkad_krb5 keys to test confdirs
We currently use afstest_BuildTestConfig() for a variety of tests to
generate a config directory, which adds a single-DES key to the KeyFile
(unless 'skipkeys' is set). But this doesn't add any non-DES keys to the
dir's KeyFileExt, and so does not exercise the more common code paths
involving non-DES encryption.

One effect of this is that if someone does a build with the wrong
endianness, our crypto routines from hcrypto may be broken, but none of
our tests will catch it (since none of them exercise hcrypto code
paths). For example, if someone runs a build on Solaris amd64 like so:

    $ ./configure --enable-bigendian

All of our tests will pass, but trying to use any servers or tools with
authenticated connections will rudely exit early on with a confusing
error, for example:

    $ vos status fs1.example.com -localauth
    vos: wrong answer

This odd message comes from check_hmac() in hcrypto as a result of a
failed sanity check, presumably because of endian-specific code in
hcrypto routines.

To catch this problem during our unit tests (and generally improve our
test coverage), add non-DES keys to the test config dir when we generate
keys. To allow for testing of code paths involving the old single-DES
keys, introduce the 'onlydeskeys' field in struct afstest_configinfo,
which allows the caller to only allow single-DES keys in the conf dir.

Change volser/vos-t to run its tests once with 'onlydeskeys' set, and
once with the default options, so volser/vos-t runs its tests with DES
and non-DES keys.

With this commit, building with the wrong endianness on Solaris now
triggers the same 'wrong answer' problem during unit tests, resulting in
at least some test failures:

    $ make check
[...]
    Failed Set                 Fail/Total (%) Skip Stat  Failing Tests
    -------------------------- -------------- ---- ----  ------------------------
    auth/superuser               26/63    41%    0    0  38-63
    volser/vos                   12/12   100%    0  255  aborted

    Aborted 1 test set, passed 5860/5898 tests.

Remove the afstest_AddDESKeyFile() function, and use a new static
function add_test_keys() instead (afstest_AddDESKeyFile has no external
callers). We use the same DES key as before, except the parity bits have
been fixed instead of calling DES_set_odd_parity() on the key.

Change-Id: Icec6145c2da0dda0562ace422a6f273930e2f9cf
Reviewed-on: https://gerrit.openafs.org/16399
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>