cf: Use AC_C_BIGENDIAN for bigendian check
Currently in OPENAFS_CHECK_BIGENDIAN, on some platforms we build and run
a small program to detect endianness (introduced in commit aaca2cacd9
(introduce-autoconf-for-openafs-20010628)). This test program is pretty
old code, lacking a return type for 'main' and lacking proper includes
for a prototype for exit(), which cause warnings on most modern
compilers.

If we're using gcc 14, these warnings cause the test to fail to build.
The way the test is structured, failing to build the test gets
interpreted as a "bigendian" result:

    configure:11245: checking whether byte ordering is bigendian
    configure:11294: gcc -o conftest     conftest.c  >&5
    conftest.c:38:1: error: return type defaults to 'int' [-Wimplicit-int]
       38 | main () {
          | ^~~~
    conftest.c: In function 'main':
    conftest.c:46:5: error: implicit declaration of function 'exit' [-Wimplicit-function-declaration]
       46 |     exit (u.c[sizeof (long) - 1] == 1);
          |     ^~~~
    conftest.c:1:1: note: include '<stdlib.h>' or provide a declaration of 'exit'
      +++ |+#include <stdlib.h>
        1 | /* confdefs.h */
    conftest.c:46:5: warning: incompatible implicit declaration of built-in function 'exit' [-Wbuiltin-declaration-mismatch]
       46 |     exit (u.c[sizeof (long) - 1] == 1);
          |     ^~~~
    conftest.c:46:5: note: include '<stdlib.h>' or provide a declaration of 'exit'
    configure:11294: $? = 1
    configure: program exited with status 1
[...]
    configure:11307: result: yes

On many platforms, such as Linux, this doesn't matter, because we infer
the endianness via the BYTE_ORDER constant (the "whether byte order is
known at compile time" configure test), and never build or run this
program. But on other platforms, such as Solaris, we don't have the
BYTE_ORDER constant, and use this test program instead.

We cannot currently build with gcc on Solaris for a variety of other
reasons, so usually this is a not a problem. But this test is clearly
rather fragile, and could become a problem in the future or on other
platforms. Before commit dba6ec9548 (cf: Set CC before calling
AC_PROG_CC), this problem can easily happen on Solaris, since CC may be
set to gcc during this endian test, before changing to the Solaris
compiler later during the 'configure' run.

While we could simply fix the test program to avoid warnings, this same
problem could easily arise if this test program fails to build for any
other reason. So to avoid this issue, get rid of the small test program
and use the AC_C_BIGENDIAN macro instead; a macro specifically for
checking endianness that comes with autoconf.

Change all users of openafs_cv_c_bigendian to use ac_cv_c_bigendian
instead, so our --enable-bigendian/--enable-littleendian options skip
the AC_C_BIGENDIAN test.

This commit keeps the existing logic for the ENDIANESS_IN_SYS_PARAM_H
define and the --enable-bigendian/--enable-littleendian options. Maybe
those could be removed in the future to be more normal about our
AC_C_BIGENDIAN usage, but for now this commit leaves those as-is so
existing code using those does not need to change.

Change-Id: I06fb4a6592ad64e93b4c9a675deefe8043d35da3
Reviewed-on: https://gerrit.openafs.org/16400
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>