rxkad/test: Remove CFLAGS from Makefile
Building the rpms under RHEL10, the link step for
src/rxkad/test/th_stress fails with:

/usr/bin/ld: th_stress.o: relocation R_X86_64_32 against `.rodata.str1.1' can not be used when making a PIE object; recompile with -fPIE
/usr/bin/ld: failed to set dynamic section sizes: bad value
collect2: error: ld returned 1 exit status

This error is because the CFLAGS are hardcoded within the Makefile and
do not contain CFLAGS that are injected by the rpmbuild process,
specifically in this case the -fPIE option.

The Makefile in src/rxkad/test was never updated to work with libtool
and still contains a CFLAGS assignment.

For this particular directory, the default MODULE_CFLAGS can be used, so
the unnecessary CFLAGS assignment can simply be removed.

Change-Id: I3fe907d11a53b398abefc6d64db3f9097d61acbc
Reviewed-on: https://gerrit.openafs.org/16404
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>