Linux: Refactor afs_linux_write_end()/begin()
Refactor afs_linux_write_end() and afs_linux_write_begin() to separate
the folio and non-folio related code.  This allows reducing some of the
preprocessor statements that are within the body of the functions.

The non-folio related code should be stable, while the folio related
code will see changes in the near future.

There are no intended functional changes in this commit. The new folio
variant of afs_linux_write_begin() now avoids setting the *foliop output
argument on error to be safe, but this should not cause any visible
change in behavior.

Change-Id: I1d167f9323165ffe203a022723320a8825e7735b
Reviewed-on: https://gerrit.openafs.org/16387
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>