linux: Make iops mkdir return a struct dentry *
The Linux 6.15 commit:
'Change inode_operations.mkdir to return struct dentry *' (88d5baf69082)
changed the signature for the return value for mkdir member of the
inope_operations structure.

The new return value for mkdir needs to be as follows:

    NULL if the directory was created and no other dentry was used
    ERR_PTR() if an error occured
    non-NULL pointer to a different dentry was used.

OpenAFS does not use a different dentry, so returning NULL or ERR_PTR
is sufficient.

Introduce a new autoconf check to determine if the inode_ops.mkdir needs
to return a pointer to a dentry.

Update afs_linux_mkdir to use ERR_PTR() on the return value.

Change-Id: I3ee3ca4ae515dcfad4075903dfcb3a37cef816f6
Reviewed-on: https://gerrit.openafs.org/16373
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>