cf: check for dentry flag macros/enums
Linux 6.15 commit:
    dcache: convert dentry flag macros to enum (b2b4483b5d05)
converted the dentry flag macros to an enum.

Because these flags are now enums, a preprocessor #ifdef can no longer
be used to test to see if the symbol is present, so a different form
of test is needed.

Introduce a new autoconf test, AC_CHECK_LINUX_SYMBOL that takes a
symbol and an include and tries to assign the symbol to an integer
variable.

Add a new autoconf file, 'linux-kernel-symbol.m4' to hold the individual
tests.

Use the new AC_CHECK_LINUX_SYMBOL to test for the DCACHE_NEED_AUTOMOUNT
and DCACHE_NFSFS_RENAMED flags.

Change-Id: Id03e6434e47df8bfb52c786ddd25ff1dac56ba63
Reviewed-on: https://gerrit.openafs.org/16377
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>