DARWIN: Check that privhelper paths start with /
In GetFullPath(), we assume that the given filename starts with a "/"
(since we concatenate it with CONFIG_PREFIX), but don't check that it
does.

Our check_filename() call currently prevents us from using any path that
doesn't start with a "/", since all paths in the hard-coded allowlist.
But just in case future commits accidentally add a non-absolute path to
the allowlist, add a check in GetFullPath() directly, so we make sure
that all paths are under CONFIG_PREFIX.

Change-Id: I224969fa889ab1aa5dabb03ff3ef0764138e5bdf
Reviewed-on: https://gerrit.openafs.org/16415
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Ganesh G. Chaudhari <gangovind@in.ibm.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>