libadmin: Use correct strlcpy() length in util_CMLocalCell()
Commit 0d6b9defb36c (libadmin: Let xdr allocate rpc output strings)
replaced a number of strcpy() with strlcpy().  Unfortunately, the size
for the strlcpy() in util_CMLocalCell() references the size of the
pointer, not the size of the cellName buffer.

This causes the 'cm_local_cell' utility to fail for cellnames of 8 or
more characters (the size of a pointer):

    $ src/libadmin/samples/cm_local_cell mvs1 7001
    util_CMLocalCell, status 21776  (ADMRPCTOOBIG)

Since util_CMLocalCell() is defined as ADMINAPI, we are not free to change
its signature.  Instead, assume the caller will provide a buffer of the
correct length UTIL_MAX_CELL_NAME_LEN.

Note: Based on the pattern of the other libadmin typedefs, it seems that
afs_CMCellName_p was intended to define a pointer to afs_CMCellName_t.
If that was the case, we could have used sizeof(*cellName) instead of
the constant UTIL_MAX_CELL_NAME_LEN.  Unfortunately, it is misdefined
and is only of type 'char*', not 'char[UTIL_MAX_CELL_NAME_LEN]*'.

Change-Id: Ic03c52db500f51f875616e2ea01c1a47e545ea7d
Reviewed-on: https://gerrit.openafs.org/16434
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>