afs: Add afs_PutCell() call in afs_lookup()
Commit d8237165f73f (afsd -dynroot-sparse mode for hushed cells) added
an afs_GetCellByName() call to afs_lookup(), but didn't include the
corresponding afs_PutCell() call to release the reference to the cell.

Currently, afs_PutCell() doesn't actually do anything because we don't
do any refcounting for cell structs. But most or all callers still try
to afs_PutCell() a cell when done with it, to try to make it more
practical to change our refcounting or locking scheme in the future (if
we ever do).

So to just try to keep the code consistent with other callers, add the
corresponding afs_PutCell() here.

Change-Id: I75cef78265df74dc6597a8cc81991ac55356a21a
Reviewed-on: https://gerrit.openafs.org/16486
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>