afs: Ensure afs_PutCell() is called in PGetWSCell()
Commit 718f85a8b69a (Fix pioctl input and output handling) added an
early return the PGetWSCell() pioctl function to return EINVAL when
output buffer cannot be set.  The early return was added after the cell
was acquired by afs_GetPrimaryCell(), so the corresponding afs_PutCell()
call to release the reference to the cell is now missing.

The missing afs_PutCell() is currently not an issue, since the
afs_PutCell() is a no-op at this time because we do not do any
refcounting for cell structs. However, this is confusing and will be an
issue if refcounting or locking is added to the cell structs in the
future.

Clean up the error handling in PGetWSCell() to ensure we release
resources before returning.

Change-Id: I474844606ee5c2d8ff44328db93cba098171f8f0
Reviewed-on: https://gerrit.openafs.org/16488
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>