util: Build tests with XLIBS
Commit 738a8e5c5c04 (util: Build src/util/test) included support for
always building dirpath_test.

However on Solaris 11.3 the build still fails:

  Undefined                       first referenced
   symbol                             in file
  gethostbyaddr                       ../util.a(hostparse.o)
  gethostbyname                       ../util.a(hostparse.o)
  socket
  /export/home/mvitale/src/sna-openafs-master/lib/librokenafs.so
  ld: fatal: symbol referencing errors
  *** Error code 2

This is because -lresolv has not been included. (This is not required on
current Solaris 11.4 because resolver symbols are now included in libc).

Add XLIBS as appropriate to repair the build errors on Solaris 11.3.
This is superfluous for Solaris 11.4 but causes no harm.

Change-Id: I177a9b8dc08877eb97f4281d07e4ebd304003c23
Reviewed-on: https://gerrit.openafs.org/16491
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>