Linux: Use __readahead_folio
The linux 6.16 commit:
  'filemap: remove readahead_page()' (c09b997342bcd)

removed the readahead_page() function.  As a replacement, the function
__readahead_folio() is functionally equivalent, except it returns a
folio instead of a page.

The Linux 5.17 commit:
  'mm/filemap: Add readahead_folio()' (9bf70167e3c61)

added __readahead_folio() and updated readahead_page() to be just a
wrapper for __readahead_folio().  The function readahead_folio() is also
just a wrapper around __readahead_folio(), however readahead_folio() is
not a direct replacement for readahead_page() as it performs a folio_put
before returning the folio.  Using __readahead_folio() allows the same
logic to be used for pages or folios.

Add an autoconf test for __readahead_folio.

Create a compatibility wrapper, afs_readahead_folio, for
__readahead_folio() / readahead_page().

Add compatibility wrappers:
    afs_FolioLocked for folio_test_locked/PageLocked
    afs_unlock_folio for folio_unlock/unlock_page.

These compatibility wrappers are for when code is working with folios
but are referencing the head page of a folio instead of referencing the
folio directly.

The Linux functions: folio_test_locked, page_folio, folio_unlock,
folio_page were introduced in 5.16.

Note: With this change, on Linux 5.17 and later the build will use
__readahead_folio instead of readahead_page.  However the actual
behavior will be the same since readahead_page was nothing more than a
wrapper for __readahead_folio.

Add curly braces to near by if statements.

Change-Id: I49acd486d604bd1110fea0a8289bdd7d1e790f4f
Reviewed-on: https://gerrit.openafs.org/16438
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>