util: Explicitly 'make all' in src/util/test
Since commit 738a8e5c5c (util: Build src/util/test), we run 'make' (with
no target) in src/util/test to make the test programs. On some platforms
(AIX, FreeBSD), this results in 'make' trying to build '%.c', which of
course we cannot do, since that's a pattern rule, and so 'make' fails:

             /bin/sh ../../libtool --quiet --mode=link --tag=CC   xlc_r   -O   -o libafsrpc_util.la base64.lo
            cd test && make
    make: 1254-002 Cannot find a rule to create target %.c from dependencies.
    Stop.
    make: 1254-004 The error code from the last command is 2.

To fix this, run 'make all' explicitly, to make all of the targets in
src/util/test.

Change-Id: Ieab3ea4152b442ccacaf9a7b5c2b8c0105364bf3
Reviewed-on: https://gerrit.openafs.org/16497
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>