util: Check for overflow and underflow in util_GetInt32 and friends
The utility functions, util_GetInt32(), util_GetUInt32(),
util_GetInt64() and util_GetUInt64() handle the parsing of strings into
numeric values.  However the parsing does not handle boundary or
overflow/underflow conditions.

For example, calling GetInt32() with the string "1111111111111111111111"
returns -1773790777.  This is because the parsing multiplies an
accumulator value by 10 and simply adds the next digit to the
accumulator, but it never checks to see if the accumulator value
overflowed or underflowed.  This parsing behavior is the same for the
three other functions.

Update the parsing to ensure the accumulator stays within the expected
range:

    util_GetInt32   -2147483648..2147483647
    util_GetUint32  0..4294967295
    util_GetInt64   -9223372036854775808..9223372036854775807
    util_GetUint32  0..18446744073709551615

With this change, these functions now return an error if we exceed the
expected range, instead of returning unexpected numbers. This can cause
some commands to fail where they previously succeeded (for example, 'pts
creategroup -id 1111111111111111111111'), but the previous results were
confusing and almost certainly not intended, so throwing an error is an
improvement.

Change-Id: Ie3312f975e096f876eff45f0677985d25eac292e
Reviewed-on: https://gerrit.openafs.org/16548
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: Michael Meffie <mmeffie@sinenomine.net>