venus: Build all venus programs by default
The cacheout, gcpags, twiddle, and whatfid programs in src/venus are not
built by default, since they are considered developer programs.  Some of
them have bit rotted and no longer build.

Add the PROGRAMS macro to the venus Makefile, which contains the list of
the programs to be built. The cacheout, gcpags, twiddle, and whatfid
programs are built but not installed at this time.

Over the years, targets were added in random order, making the makefile
harder to read.  Clean up the makefile by putting the link and compile
targets in separate sections and put the targets in sorted order to make
this easier to maintain.

Update the .gitignore as well to be sure we ignore all of the programs
that are built.

Change-Id: I4dfac5149bedb3ed4d3e402f0bd88c7086b44e95
Reviewed-on: https://gerrit.openafs.org/16659
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>