DARWIN: Handle ',' and '/' openssl x509 delims
Our logic for extracting MACOS_TEAM_ID from our application signing cert
assumes that the output from 'openssl x509 -noout -subject' looks like
this:

    $ security find-certificate -c xxxx -p | openssl x509 -noout -subject
    subject=UID=SKMME9E2Y8, CN=Developer ID Application: Org (SKMME9E2Y8), OU=SKMME9E2Y8, O=org, C=US

That is, the components are delimited by commas.

But current macOS uses an 'openssl' from LibreSSL, and the output looks
like this:

    $ security find-certificate -c xxxx -p | openssl x509 -noout -subject
    subject= /UID=SKMME9E2Y8/CN=Developer ID Application: Org (SKMME9E2Y8)/OU=SKMME9E2Y8/O=Org/C=US

That is, the components are delimited by slashes. So currently, this
build logic extracts the MACOS_TEAM_ID as something like
"SKMME9E2Y8/O=Org/C=US", which is incorrect. Nothing notices the
error, and the build proceeds, eventually building binaries with the
garbage team id, and so our XPC security checks may always fail for the
menu helper and backgrounder.

To fix this, try to accommodate both output styles, delimited by either
commas or slashes. Add a check to see if the extracted MACOS_TEAM_ID
contains extra components, and fail if so, since that suggests we didn't
extract it properly.

Change-Id: Iee1980a7985870679d4d9255600bb2f81f1b90ba
Reviewed-on: https://gerrit.openafs.org/16517
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>