DARWIN: Avoid double AuthorizationFree()
Since its introduction with commit 694c5ca3f400 (OSX Preference Pane
Source Import), the autorize() function of the OpenAFS prefpane obtains
a global authorizationRef, and deautorize() frees it.  Later, with
commit 120871f03f38 (DARWIN: Add 'privhelper' tool for PrefPane), a
state flag isAuthorizationRefOwned was introduced.

However, if autorize() encounters an error, (e.g., the user cancels the
password dialog prompt for the AFSMenu option "Startup AFS" or "Stop
AFS"), it may also free authorizationRef.  But it does not set
authorizationRef to nil, nor does it update the isAuthorizationRefOwned
flag.  Therefore, when deautorize() is next called, it will attempt to
free authorizationRef again.  This causes the legacyLoader (which is
hosting the prefpane code) to SIGABORT.

Instead, after AuthorizationFree, ensure that we set authorizationRef to
nil and isAuthorizationRefOwned to NO.

Change-Id: I6dfc5ddb809db14c14adb4705c66b6b616cb5e4f
Reviewed-on: https://gerrit.openafs.org/16523
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>