DARWIN: Release cacheInfoStrData in readCacheInfo
Since the introduction of the OpenAFS prefpane (commit 694c5ca3f400 'OSX
Preference Pane Source Import'), the function readCacheInfo() has
allocated a temporary local storage to parse the cacheinfo file, but the
release of this storage has been commented out.  This results in a
memory leak proportional to the size of the cacheinfo file on every call
to readCacheInfo().

Uncomment the release to fix the memory leak.

Discovered via the macOS 'leaks' utility.

Change-Id: I8cc25006c959f8e98a653c3c382cb3800f7cf91b
Reviewed-on: https://gerrit.openafs.org/16525
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>