DARWIN: Ensure 'make packages' fails if codesign fails
Although pkgbuild.sh has 'set -e' in force, the following 'find' command
does not stop the build if 'codesign' fails for any reason:

    find "$PKGROOT" -type f -exec codesign --verbose --force \
           --timestamp --sign "$APP_KEY" $CODESIGN_OPTS {} \;

This is normal documented behavior for 'find'; failed predicates do not
influence the return code of 'find' itself.

Modify the logic in pkgbuild.sh.in to ensure the build will stop if
codesign fails for any file.

Discovered during a network outage when codesign failed with:
  'The timestamp service is not available.'

While here, clarify the comment for the other codesign stanza.

Change-Id: If46cb04eed6d091dc3ac9f3ca7b6b17d446b041e
Reviewed-on: https://gerrit.openafs.org/16532
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>