util: Fix infinite loop in util_GetInt64(), util_GetUInt64()
The util_GetInt64() and util_GetUint64() functions will loop
infinitely when the input string has leading spaces or tabs.

This bug was fixed in the 32 bit variants in commit 69bab4cc6ee6
(util-volparse-20090122).  The 64 bit functions were introduced after
that fix in commit 4e3d3d50496c (afsio is a command to pipe data into or
out of afs files), presumably based on the unfixed versions on an older
branch.

Replace the while() loops in the 64 bit functions to match what is done
in the 32 bit variants to fix this bug and make the function
implementation more consistent.

Currently the only in-tree callers of util_GetInt64 is `afsio` and the
tabular_output.c module. The tabular_output.c module is not used at this
time.  The util_GetUInt64 has no in-tree callers at this time.

Change-Id: I32528806a60bb0ce3f8d0a98060fed1dd1b1a4fa
Reviewed-on: https://gerrit.openafs.org/16547
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: Michael Meffie <mmeffie@sinenomine.net>