bos: Check util_GetInt32 return code in bos setrestrict
The `bos setrestrict -mode` option value is converted to an integer from
the input string with the util_GetInt32() function, but the return code
is not checked. Change SetRestrict() to print an error message and exit
if util_GetInt32() returns an error, so we don't send 4 bytes of
uninitialized data to the bosserver when we're given a bad -mode value.

Change-Id: I5891698d1371f8a94bc39eaf0b49e28ce7b4ec84
Reviewed-on: https://gerrit.openafs.org/16551
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>