lwp: Rename timer.h to afs/timer.h
Since OpenAFS 1.0, we have installed a /usr/include/timer.h. Move it to
/usr/include/afs/timer.h to avoid possible conflicts and make it
clearer this is an OpenAFS-related file. While there are currently no
known conflicting header files from other projects (though there is a
sys/timer.h on AIX), this is a very generic-sounding name, so move it
before it becomes a problem.

It seems unlikely there are any users of timer.h outside of OpenAFS;
there aren't even any users in OpenAFS outside of src/lwp. But keep
installing it for now, in case it is needed.

Change-Id: Ia3a9d6de828814e260d255e107031a145b0ebda4
Reviewed-on: https://gerrit.openafs.org/16557
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>