Linux: Use struct kiocb * for aops write_begin/end
The Linux 6.17 commit:
    fs: change write_begin/write_end interface to take struct kiocb *
    (e9d8e2bf23206)
changed the address_space_operations members write_begin and write_end
to use a 'struct kiocb *' instead of a 'struct file *' as the first
parameter.

The passed kiocb structure contains a pointer (ki_filp) to the file
structure that is associated with the operation.

Update the afs_linux_write_begin() and afs_linux_write_end() to accept
a 'struct kiocb *' instead of a 'struct file *', and obtain the file
pointer directly from kiocb.

Add an autoconf test to determine if aops->write_begin()/aops->write_end
uses a struct kiocb * as the first parameter.

Change-Id: I53878ca8f671d05e318e783895d9e2fa280ac359
Reviewed-on: https://gerrit.openafs.org/16558
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>