linux: remove implied def HAVE_LINUX_FILEMAP_GET_FOLIO
Commit:
  Linux: Use struct kiocb * for aops write_begin/end (1ad827ded0ed9)
added 2 implied defines in osi_vnodeops, one for
HAVE_LINUX_WRITE_BEGIN_END_FOLIO and one for HAVE_LINUX_FILEMAP_GET_FOLIO.
The HAVE_LINUX_FILEMAP_GET_FOLIO was not needed, and causes a build
failure:

  osi_vnodeops.c:4056:10: error: "HAVE_LINUX_FILEMAP_GET_FOLIO" redefined [-Werror]
   4056 | # define HAVE_LINUX_FILEMAP_GET_FOLIO
        |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~

This is because the configure test for __filemap_get_folio() succeeds
(linux 6.15 and later) and the test for write_begin/end functions take a
kiocb also succeeds (6.17 or later).

Remove the unnecessary define.

Change-Id: I9a5eadeae911ded80698d7246a08c53028f937a1
Reviewed-on: https://gerrit.openafs.org/16589
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>